[FE-201] refactor: 마이 레코드 페이지 api 분리
