+++ src/apis/record.ts
@@ -1,14 +1,5 @@
 import { parentCategoryID } from './../types/category'
-import { AxiosResponse } from 'axios'
-import {
-  IMemoryRecordList,
-  IMyRecordRequestParam,
-  IRecordByDateList,
-} from 'types/recordData'
 import { baseInstance } from './instance'
-
-const MEMORY_RECORD_SIZE = 7
-const MEMORY_COMMENT_SIZE = 5
 
 export const getCategory = async (categoryId?: parentCategoryID) => {
   return await baseInstance.get('/record/category', {
@@ -29,18 +20,6 @@ export const getRecord = async (recordId: string | undefined) => {
   }
 }
 
-export const getMemoryRecord = (
-  pageParam: number
-): Promise<AxiosResponse<IMemoryRecordList>> => {
-  return baseInstance.get(`/record/memory`, {
-    params: {
-      memoryRecordPage: pageParam,
-      memoryRecordSize: MEMORY_RECORD_SIZE,
-      sizeOfCommentPerRecord: MEMORY_COMMENT_SIZE,
-    },
-  })
-}
-
 export const deleteRecord = async (recordId: string | undefined) => {
   if (recordId) {
     const res = await baseInstance.delete(`/record/${recordId}`)
@@ -54,20 +33,6 @@ export const modifyRecord = async (
 ) => {
   return baseInstance.put(`/record/${recordId}`, data, {
     headers: { 'Content-Type': 'multipart/form-data' },
-  })
-}
-
-export const getRecordByDate = ({
-  date,
-  page,
-  size,
-}: IMyRecordRequestParam): Promise<AxiosResponse<IRecordByDateList>> => {
-  return baseInstance.get(`/record`, {
-    params: {
-      date,
-      page,
-      size,
-    },
   })
 }
