+++ src/react-query/hooks/useRecordByDate.ts
@@ -1,6 +1,6 @@
 import { QUERY_KEYS } from '@react-query/queryKeys'
 import { useQuery } from '@tanstack/react-query'
-import { getRecordByDate } from '@apis/record'
+import { getRecordByDate } from '@apis/myRecord'
 import { getFormattedDate } from '@utils/getFormattedDate'
 import { getMonthYearDetail } from '@pages/MyRecord/Calendar/getCalendarDetail'
 import { useState } from 'react'
